requirements.txt
tensorflow
ultralytics
matplotlib
pyyaml
numpy